+++ CHANGELOG.md
@@ -9,12 +9,14 @@ Added:
     - HTML5: Button.
     - DaisyUI: Hero, HeroContent.
     - DaisyUI: Indicator, IndicatorItem.
+    - DaisyUI: Stack.
     - Examples: Layout.
 
 Changed:
     - Tailwind Plugins: forms, typography, aspect-ratio, line-clamp.
     - DaisyUI: Button, support for text label in constructor.
     - HTML5: Span is subclass of TextContentComponent.
+    - Component: `Component` go all features from `TextContentComponent`.
 
 Removed:
     - Examples: Artboard, Divider, Footer.